+++ application-module/src/main/kotlin/com/reservation/user/history/access/port/input/CreateUserAccessHistoriesCommand.kt
@@ -1,7 +1,7 @@
 package com.reservation.user.history.access.port.input
 
 import com.reservation.enumeration.AccessStatus
-import com.reservation.shared.user.LoginId
+import com.reservation.user.shared.LoginId
 import com.reservation.user.history.access.port.output.CreateUserAccessHistories.CreateUserHistoryInquiry
 import java.time.LocalDateTime
