+++ application-module/src/main/kotlin/com/reservation/user/self/port/output/LoadGeneralUser.kt
@@ -1,8 +1,8 @@
 package com.reservation.user.self.port.output
 
-import com.reservation.shared.user.LoginId
-import com.reservation.shared.user.Password
-import com.reservation.shared.user.PersonalAttributes
+import com.reservation.user.shared.LoginId
+import com.reservation.user.shared.Password
+import com.reservation.user.shared.PersonalAttributes
 import com.reservation.user.self.User
 import java.time.LocalDateTime
