+++ application-module/src/test/kotlin/com/reservation/user/self/usecase/FindGeneralUserPasswordUseCaseTest.kt
@@ -5,9 +5,9 @@ import com.reservation.common.exceptions.NoSuchPersistedElementException
 import com.reservation.exceptions.InvalidSituationException
 import com.reservation.fixture.CommonlyUsedArbitraries
 import com.reservation.fixture.FixtureMonkeyFactory
-import com.reservation.shared.user.LoginId
-import com.reservation.shared.user.Password
-import com.reservation.shared.user.PersonalAttributes
+import com.reservation.user.shared.LoginId
+import com.reservation.user.shared.Password
+import com.reservation.user.shared.PersonalAttributes
 import com.reservation.user.self.User
 import com.reservation.user.self.port.input.FindGeneralUserPasswordCommand.FindGeneralUserPasswordCommandDto
 import com.reservation.user.self.port.output.LoadGeneralUserByLoginIdAndEmail
