+++ core-module/src/main/kotlin/com/reservation/resign/self/ResignedUser.kt
@@ -1,6 +1,6 @@
 package com.reservation.resign.self
 
-import com.reservation.shared.user.LoginId
+import com.reservation.user.shared.LoginId
 import java.time.LocalDateTime
 
 class ResignedUser(
