+++ core-module/src/main/kotlin/com/reservation/user/owner/service/CreateSellerUserService.kt
@@ -1,8 +1,8 @@
 package com.reservation.user.owner.service
 
-import com.reservation.shared.user.LoginId
-import com.reservation.shared.user.Password
-import com.reservation.shared.user.PersonalAttributes
+import com.reservation.user.shared.LoginId
+import com.reservation.user.shared.Password
+import com.reservation.user.shared.PersonalAttributes
 import com.reservation.user.owner.RestaurantOwner
 import com.reservation.user.policy.formats.CreateUserFormats
 import com.reservation.utilities.encrypt.password.PasswordEncoderUtility
