+++ core-module/src/main/kotlin/com/reservation/user/self/User.kt
@@ -1,12 +1,13 @@
 package com.reservation.user.self
 
 import com.reservation.enumeration.Role
+import com.reservation.enumeration.Role.USER
 import com.reservation.resign.self.EncryptedAttributes
 import com.reservation.resign.self.ResignedUser
-import com.reservation.shared.user.LoginId
-import com.reservation.shared.user.Password
-import com.reservation.shared.user.PersonalAttributes
-import com.reservation.shared.user.UserAttribute
+import com.reservation.user.shared.LoginId
+import com.reservation.user.shared.Password
+import com.reservation.user.shared.PersonalAttributes
+import com.reservation.user.shared.UserAttribute
 import com.reservation.user.common.exceptions.ResignWithoutIdException
 import com.reservation.user.policy.formats.ServiceUser
 import java.time.LocalDateTime
@@ -21,7 +22,7 @@ class User(
     private var personalAttributes: PersonalAttributes,
     nickname: String,
 ) : ServiceUser {
-    private var userAttributes: UserAttribute = UserAttribute(nickname, Role.USER)
+    private var userAttributes: UserAttribute = UserAttribute(nickname, USER)
 
     override val identifier: String?
         get() = id
