+++ core-module/src/main/kotlin/com/reservation/user/shared/LockState.kt
@@ -1,6 +1,8 @@
-package com.reservation.shared.user
+package com.reservation.user.shared
 
 import com.reservation.enumeration.UserStatus
+import com.reservation.enumeration.UserStatus.ACTIVATED
+import com.reservation.enumeration.UserStatus.DEACTIVATED
 import java.time.LocalDateTime
 import java.time.temporal.TemporalUnit
 
@@ -20,9 +22,9 @@ data class LockState(
         unit: TemporalUnit,
     ): Boolean = lockedDateTime?.plus(interval, unit)?.isBefore(LocalDateTime.now()) ?: true
 
-    fun deactivate(): LockState = LockState(failCount, LocalDateTime.now(), UserStatus.DEACTIVATED)
+    fun deactivate(): LockState = LockState(failCount, LocalDateTime.now(), DEACTIVATED)
 
-    fun activate(): LockState = LockState(0, null, UserStatus.ACTIVATED)
+    fun activate(): LockState = LockState(0, null, ACTIVATED)
 
     fun addFailureCount(): LockState = LockState(failCount + 1, lockedDateTime, userStatus)
 }
