+++ core-module/src/main/kotlin/com/reservation/user/shared/Password.kt
@@ -1,4 +1,4 @@
-package com.reservation.shared.user
+package com.reservation.user.shared
 
 import java.time.LocalDateTime
