+++ core-module/src/main/kotlin/com/reservation/user/shared/PersonalAttributes.kt
@@ -1,4 +1,4 @@
-package com.reservation.shared.user
+package com.reservation.user.shared
 
 data class PersonalAttributes(
     val email: String,
