+++ core-module/src/main/kotlin/com/reservation/user/shared/UserAttribute.kt
@@ -1,4 +1,4 @@
-package com.reservation.shared.user
+package com.reservation.user.shared
 
 import com.reservation.enumeration.Role
