+++ core-module/src/test/kotlin/com/reservation/resign/ResignUserServiceTest.kt
@@ -2,9 +2,9 @@ package com.reservation.resign
 
 import com.reservation.fixture.CommonlyUsedArbitraries
 import com.reservation.resign.self.service.ResignUserService
-import com.reservation.shared.user.LoginId
-import com.reservation.shared.user.Password
-import com.reservation.shared.user.PersonalAttributes
+import com.reservation.user.shared.LoginId
+import com.reservation.user.shared.Password
+import com.reservation.user.shared.PersonalAttributes
 import com.reservation.user.self.User
 import com.reservation.utilities.encrypt.bidirectional.BidirectionalEncryptUtility
 import com.reservation.utilities.generator.uuid.UuidGenerator
