+++ core-module/src/test/kotlin/com/reservation/user/general/ChangeGeneralUserPasswordServiceFailureTest.kt
@@ -3,9 +3,9 @@ package com.reservation.user.general
 import com.navercorp.fixturemonkey.kotlin.giveMeOne
 import com.reservation.fixture.CommonlyUsedArbitraries
 import com.reservation.fixture.FixtureMonkeyFactory
-import com.reservation.shared.user.LoginId
-import com.reservation.shared.user.Password
-import com.reservation.shared.user.PersonalAttributes
+import com.reservation.user.shared.LoginId
+import com.reservation.user.shared.Password
+import com.reservation.user.shared.PersonalAttributes
 import com.reservation.user.common.exceptions.InvalidateUserElementException
 import com.reservation.user.common.exceptions.UseSamePasswordAsBeforeException
 import com.reservation.user.self.User
